test: 테스트 Promotion 테스트 파일 분리 구현 Promotion의 메소드 테스트중 의존성이 필요한 경우가 발생해 2개의 파일로 분리, PromotionTest: 의존성이 포함된 테스트. PromotionMockTest: mock 함수를 이용한 테스트.
